Rename CreateGroup command folder to Group.
