+++ src/StudentToolkit/WpfCore/Commands/Presentation/Group/ShowCreateStudentDialogCommand.cs
@@ -1,6 +1,6 @@
 ﻿using StudentToolkit.MVVM.Validation.CreateGroup;
 
-namespace StudentToolkit.WpfCore.Commands.Presentation.CreateGroup;
+namespace StudentToolkit.WpfCore.Commands.Presentation.Group;
 
 public sealed class ShowCreateStudentDialogCommand : Command
 {
